Change Dialog return to ResponseType
While the Dialog return type is noted as a gint in the documentation, it
is noted there that the return type is either GTK_RESPONSE_NONE if the
dialog was destroyed, or the response type of the button that was
clicked. Since this library has created a separate ResponseType type, it
makes sense to cast the integer to a ResponseType in the library instead
of requiring the developer to make the cast in each of their dialog
uses.